Disable Assembly Tests for 1.8.1 release (#5875)
## What is the goal of this PR?
Due to dependency issues between mixed 2.0 and non-2.0 dependencies, we were unable to re-enable the client-java based distribution and assembly tests. The solution to this should be to include `client-java` as an artifact (via maven resolution) from our snapshot repositories for the test. However, in the interest of getting the release ready and moving forward, we have temporarily disabled assembly tests using client-java. They should be brought back in Grakn 2.0

## What are the changes implemented in this PR?
* Disable assembly tests for windows, mac, linux and docker as they utilised client-java