Refactor out BatchSize.ALL and Transaction.Iter.Req.Options (#32)
## What is the goal of this PR?

To make the RPC interface for queries simpler

## What are the changes implemented in this PR?

Refactored out BatchSize.ALL. It was previously converted to Integer.MAX_VALUE on the backend. Now, if you really want to pass in a very high batch size, you can, but we don't expose it as a separate field.

We then removed Transaction.Iter.Req.Options in favour of using the Options.BatchSize set at the Transaction or Session level.