Rename Schema to Encoding (#35)
## What is the goal of this PR?
To free up the name `schema` for other uses, we rename the use of the low-level byte constants from `Schema` to `Encoding`. This has the added benefit of being more "low-level", which is closer to the actual meaning of the class.

## What are the changes implemented in this PR?
* Rename `Schema` to `Encoding`, and replace variables called `XXX_SCHEMA` to `XXX_ENCODING`
* Rename methods  using `schema()` to `encoding()`

After this PR, the only references to Schema should be at the session level!