Implement Rules (#36)
## What is the goal of this PR?
Implement Rules in the schema graph, including the rule edges that are encoded as SchemaEdges. This will allow us to perform the required operations to fire relevant rules during the resolution process. Rules are NOT concepts, but live in the `concept.schema` package. We also align terminology globally by renaming `ThingGraph` to `DataGraph`

TODO: extract types from rules to write the correct rule edges. This depends on the completion on Graql as a graph and then extracting labels from the patterns.

## What are the changes implemented in this PR?
* `TypeVertex` became `SchemaVertex`, with subtypes `TypeVertex` and `RuleVertex`
* Implement `Rule`, and `RuleVertex`
* Extend Encodings to write labels, rule `when` and `then`
* Extend Concepts to write and read rules
* Simplify the generics on `Vertex`, `Graph`, and `Adjacency`, the latter which has becomes `AdjacencyDirection`
* Split schema and thing graphs fully, as users always know 
* Rename `ThingGraph` to `DataGraph`, and relevant usages
* Rename `concept.type` to `concept.schema` package, which contains `Rule` and `impl/RuleImpl`