Fix import cache bug for migration (#5884)
## What is the goal of this PR?

We fixed an issue in import where, under rare conditions, the cache would lookup an incorrect concept in the final stage of importing and cause the import to fail if this concept was not a valid attribute (or produce incorrect results if it was).

At the end of the import, we insert any ownerships that could not be inserted earlier (because one of the ends, owner or attribute, were not yet encountered in the data). During this step, we use a local cache to look up any already "live" attributes that used in the same transaction to avoid reading the same attribute through Concept API repeatedly.

The bug was that the lookup ID for missing attribute ownerships was being translated from an "original" ID to an "imported" ID before lookup, but the cache works using the "original" ID to save unnecessary translation. Fortunately, whilst the bug itself could produce incorrect results, it was unlikely to occur commonly. The translated ID was also being used to cache attributes, so any attributes loaded at this stage would be cached and looked up correctly.

The migration would fail if there was an ID collision between the "original" ID of an entity or relation and the "new" ID of an attribute. The migration would potentially succeed but with incorrect results if an attribute was cached during the phase before the final missing ownerships phase where its "original" ID collided with a "new" attribute ID during the same batch. The collisions would be uncommon in most cases, and very uncommon for the attribute case that might go undetected. Since the cache is cleared between transactions, the probability would also be limited (not scaling with data set size beyond a point.)

## What are the changes implemented in this PR?

- Fix the cache lookup to be by original ID for missing ownership insertions.