Introduce Grakn 'dev' mode backed by RocksDB JNI 'dev' version (#41)
## What is the goal of this PR?

We discovered that RocksDB was throwing a C++ assertion error on closing a database. This was highly worrying and we considered it a possible explanation of our segfault issues, which have been plaguing us while trying to test Grakn.

RocksDB (and our own Java) assertions are turned off at runtime, currently. We want to make Grakn, and especially RocksDB, more debuggable.

## What are the changes implemented in this PR?

Grakn now depends on two distinct RocksDB JNI jars: a "production" JAR, obtained from Maven's main repository, and a "development" JAR, custom-built by us and deployed to our Nexus.

But at any time, it will only ever actually import one of these JARs as a dependency.

By default, Bazel builds of Grakn will now use the development JAR, with assertions switched on. Meanwhile, when building a Grakn distribution, the number of paths that our JARs get built to is increased to 3:

- `server/lib/common` - JARs common to both debug and production mode
- `server/lib/dev` - JARs used only in debug mode
- `server/lib/prod` - JARs used only in production mode

RocksDB JNI dev goes to `dev`, RocksDB JNI (production) goes to `prod`; everything else goes to `common`.

This will allow Grakn to be run in either debug or production mode by the user, by simply running either `./grakn server start` or `./grakn server start --debug`.

## Other information

This PR also fixes a potential memory leak where RocksIterators are not always cleaned up correctly.

NOTE: We are currently unable to build a cross-platform debug RocksDB JNI JAR (see https://github.com/graknlabs/dependencies/issues/207 for details). Because of this limitation we need to make the production JAR the default used in CI and IDEs for now. But --debug will still work in distribution and can be easily toggled in an IDE.