2nd RPC Rework: Server (#59)
## What is the goal of this PR?

Reorganise and simplify the RPC package, and introduce latency-adaptive answer sending, in order to make Grakn clients easier to implement and much faster.

## What are the changes implemented in this PR?

- Update the RPC package to adhere to the latest protocol and refactor almost all its classes into clean, stateless request handlers
- Create `SessionRPC` and `TransactionRPC` as explicit state variables; their methods mirror those in `Session` and `Transaction`
- Separate the Transaction logic into `TransactionRPC` and `TransactionStream`, where `TransactionRPC` is responsible for processing and responding to queries and commits, and `TransactionStream` is responsible only for handling the opening and closing of the bidirectional stream, and the opening of the `Transaction`. This allows `TransactionRPC` itself to be fully stateless (i.e: all variables are `final`) despite the limitation that `TransactionStream` cannot be stateless because we don't have a reasonable way of sending the session ID at the time that the stream is opened.
- Add network latency compensation to streaming Transaction responses, eliminating time wasted waiting for network roundtrips
- Increase the number of answers the server precomputes to 128, allowing clients to potentially fetch whole batches immediately if they consume answers slower than the server produces them

### Latency compensation (aka smart batching)

Clients can now optionally specify their network latency when issuing a query request.

If it is a query that supports iteration (e.g. `ThingType.getInstances()`) and it has many answers then the server will, in order:
- compute and send _N_ answers, where _N_ is the batch size;
- send a `continue` signal, a special response message that instructs the client to issue a `continue` request if they want another batch;
- continue computing and sending answers for another _L_ ms, where _L_ is the network latency.

The result is that, if the client's consumption rate is high and their network latency remains stable, the server will continue sending responses to the client right up until the point where the client's next `continue` request comes in. Without latency compensation, _L_ ms would have been effectively wasted between sending the last item in the first batch, and the first item in the second batch.