Infer label hints on thing variable isa and role player constraints (#69)
## What is the goal of this PR?

Reasoner will now have the ability to infer which type labels are possible within a conjunction. The result of this calculation will be stored as a Set of Labels on the isa constraints and rolePlayers, as required.

## What are the changes implemented in this PR?

A new class, TypeInference, has been made that will carry the logic of converting a conjunction into a conjunction query that the original conjunction's types will have to satisfy. For example, for the conjunction 
```
$p isa persons, has name 'bob';
```
will return the conjunction:
```
$p sub person, owns name;
```
This is run through as a query, resulting in all the types `$p` can satisfy from the original query.

This required making typeConstraint constructors public.
TypeConstraint helper constructors also written.