Concludables (#100)
## What is the goal of this PR?

We introduce Concludables, which are the units of reasoning for Grakn's reasoner. When considering a conjunction, Grakn needs to break that conjunction down into blocks for evaluation against the `then` of the rules in the KB. To do this conjunctions are broken down into Concludables. The same applies to the conjunction in the `when` of a rule in order to support chained reasoning. The `then` of a rule is also broken down into concludables, such that these can be matched against conjunction concludables in order to find applicable rules.

There is a ranking among Concludables when breaking down a conjunction. `Relation` and `Has` take precedence over `Isa` and `Value`. This is to reduce the number of times the same rule will match against the same part of the conjunction. For example, a relation `($x, $y) isa marriage;` will create one Concludable, a `Concludable.Relation`, from a conjunction. The same Graql used in the `then` of a rule will create a `Concludable.Relation` and a `Concludable.Isa`. In this way the rule can be matched by concludable when a user (or rule) queries `match $x isa thing;` or `match ($x, $y) isa marriage;`. See `ConcludablesTest` for far more examples.

## What are the changes implemented in this PR?

Entirely introduces: 
- `Concludable`
- `ConcludableTest` - unit tests
- `Implication` - a wrapper of `Rule` for reasoner's purpposes
- `ImplicationTest` - integration tests
- `Unification` - to be fleshed out later, was deemed scope creep for this PR

Changes:
- Some changes were needed to Variables and Constraints to add helper methods.
- A rule `then` was changed from a set of Variables to a set of Constraints