Make rules immutable and remove RPC define returns (#56)
## What is the goal of this PR?
Use newest protocol to reflect: not return anything on define, and remove setters on rule when and then. 
We also correspondingly make rule vertices immutable (ie. remove `setWhen` and `setThen` on `RuleVertex`).
Finally, calling `ConceptManager.putRule` deletes and re-creates the rule if putting a rule with a prexisting label.

We also clean up the implementation putting rule edges.

## What are the changes implemented in this PR?
* Update protocol
* Remove returns for `define`
* Remove rule setters for `when` and `then` all the way down to `RuleVertex`
* Simplify putting rule hypothesis edges