Type Hinting Tests (#117)
## What is the goal of this PR?

Tests for Type Hinter. We want to be able to check that the possible types calculated are as expected. This test suite also includes several edge cases that demonstrate some counter-intuitive results. This tests both modes of Type Hinter - exhaustive tests and simple one-hop type hinting tests.

## What are the changes implemented in this PR?

Adds `TypeHinterTest.java`, and accompanying schema to help it with this.