Alpha Equivalence (#134)
## What is the goal of this PR?
Introduce Alpha Equivalence for Patterns, to facilitate finding alpha-equivalent concludables during the actor-lookup for a particular concludable.

## Requirement
This is a requirement for the reasoner actor model to terminate during reasoning cycles. During cycles we require that messages for identical Concludables are always received by the same actor. In this way, the actor can terminate the cycle if it receives a message, with an id map, that it has explored before.

Within a Concludable, alpha equivalence falls short of full graph isomorphism checking. Thus, we simply check a constraint by checking all of its variables except for the owner (checking the owner would create a cycle). To check for alpha equivalence of a Concludable therefore, we check the alpha equivalence of it's constraint's owner.

## What are the changes implemented in this PR?
An `alphaEquals` methods on Variables and Constraints. 

If alpha equivalence is true we also need a variable map between the two constraints. To facilitate this we have `AlphaEquivalence` which has a fluent interface. The AlphaEquivalence can be invalidated by any stage of the building of the alpha equivalence. If an alpha equivalence is valid valid, it will contain a `map` between the variables concerned.

We also build a `PermutationIterator` to iterate over all permutations of role players in a relation constraint.