Rearchitect schema cache by splitting schema db and data db (#136)
## What is the goal of this PR?

We have rearchitected the schema cache.

Before this PR, we keep a long-running schema transaction that is shared across data transaction. The inner transaction of the schema transaction needs to be refreshed after 10 commits, since we want to prevent the RocksDB snapshot from growing too large.

An old inner transaction that has been refreshed (ie., replaced by a new one) can only be closed when there are no more data transactions holding references to its iterators. Keeping track of live iterators are error-prone and we want to find a simpler solution.

In this PR, we split the schema and data each into its own RocksDB database. Given that schema and data are no longer bound to the same RocksDB snapshot, we no longer need to periodically refresh the schema transaction. Not having to refresh the schema transaction has simplified the architecture considerably.

## What are the changes implemented in this PR?

1. Split a Grakn database into two RocksDB databases: one for schema, one for data:
```
server/
  data/
    my_database/
      schema/
      data/
```
2. `RocksTransaction.Data` access the schema via a shared `SchemaGraph`
3. `RocksTransaction.Schema` access the data by creating its own instance of `DataGraph`
4. Extracted the storage objects into its own class `RocksStorage`