Long lived rule cache and rule validation hooks (#5923)
## What is the goal of this PR?

Avoid re-loading rules that are in the rule cache by introducing long-lived `LogicCache`. The `LogicCache` contains the prior`TypeHinterCache` and the new `RuleCache`. We also introduce a scan for retrieving all rule structures that are stored in the schema graph, and leave hooks for rule validation points. Rule validation now occurs on both rule `put` and on schema transaction `commit`

## What are the changes implemented in this PR?

* Introduce `LogicCache`, containing `RuleCache` and `TypeHinterCache`
* Introduce a scan for all rules in `SchemaGraph`. We will improve this by indexing rules by types they mention in the future, so we can look up rules that are relevant to a starting rule and the types it can infer
* Introduce points for rule validation: at `put` time, and re-validation at schema transaction commit time (in case the schema has chanced since writing the rule the first time)
* rename `Grakn.Transaction.logics()` to `Grakn.Transaction.logic()`
* Simplify `Rule` by removing `ConceptManager`