TypeHints moved from Constraint Field to Variable Field (#5936)
## What is the goal of this PR?

Moved the Type Hints (the set of possible types that a variable can be), from one of its constraints (`isaConstraint` for `ThingVariable` and `subConstraint` for `TypeVariable`) to `Variable`. `TypeHints` in `RolePlayer` remain unchanged.

We also introduce a new `isSatisfiable` flag onto `Variable`. This acts as a guarantee of the satisfiability of a `Variable`, and is important in the case of an empty `TypeHints` (which can mean either a Variable that can't be anything, or can be anything, depending upon this flag).

## What are the changes implemented in this PR?

Field `typeHints` and its accompanying methods were moved from `isaConstraint` and `subConstraint` to `Variable`. Code calling on typeHints have been changed accordingly.

Added `isSatisfiable` flag, and added some logic to changing it when TypeHints are added to and adjusted in `TypeHinter` and `Rule`.