Persist statistics snapshot into database (#5953)
## What is the goal of this PR?

To make us recalculate query plan only when needed, we keep a snapshot version of the current statistics. Currently this snapshot is not persisted into database, which means we lost the information on if we should recalculate query plan across data transactions. Therefore we need to persist this snapshot number into database.

## What are the changes implemented in this PR?

- Save statistics snapshot number into database.