Unification from Conjunction Concludables to Head Concludables (#5933)
## What is the goal of this PR?

Given a Conjunction Concludable, we can return all possible unifiers from the Conjunction Concludable to Head Concludable.

In other words, we can find every valid mapping from the named variables of a concludable from a Conjunction or a Query, to a corresponding Concludable that might appear as a Concludable to the Then of a Rule. 

## What are the changes implemented in this PR?

Implemented a `unify` method to Conjunction Concludable that takes in a Head Conjunction and returns a Stream of Unifications.