Answer State Tracking (#5957)
## What is the goal of this PR?

We improve that way that the state of answers in reasoner is tracked throughout the reasoner. This new architecture aims to make clear which variables domain an answer state is in via it's namespace, and therefore which state it can move to next.

## What are the changes implemented in this PR?
This needs to support differing methods for unified answers, mapped answers and answers for the root resolver. Most crucially that we can always expect a valid answer after a 1:1 variable mapping, but an answer may prove invalid when unified or un-unified.