Rearchitect Unifiers to include Requirements (#5977)
## What is the goal of this PR?
We modify the architecture of `Unifier` to impose `Requirements` when doing reverse unification - in lieu of query rewriting, we must filter some materialised rule answers out when they do not match the user's exact query. 
We also redesign the general flow of computing `Unifier`s in `Concludable` to have a clearer flow, using a `Unifier.Builder` to avoid needless copying. This also leads to simplifying the states in `AnswerState`.


## Unifier "Requirements"
A major decision to this point was to avoid recursive query rewriting for the time being (https://github.com/graknlabs/grakn/issues/5921). To implement correct unification, a `Unifier` may now impose `Requirements` on a reverse unification that is being performed. We may require a Variable is one of a set of `isaExplicit` (exact type of an instance), or a type in a set of `types` (exact label of a type), or a `predicate`.

__isaExplicit__ and __predicate__ requirements stem from the following case:
query: `match $p has name "john"`
rule: `when {$x isa person; $y isa first-name;} then {$x has $y}`

Here, the rule my indicate that an instance has an attribute, but the user only requests concepts with `name` attributes, that are specifically `= "john"`. So in the Unifier, we have approximately: 
`Unifier.requirements().isaExplicit()` contains `set("name", "first-name", "last-name")` 
and
`Unifier.requirements().predicates()` contains `(attr) -> attr.value().equals("john")`

__type__ requirements stem from the following case:
query: `match $r isa friendship`
rule: `when {$x isa person;} then { (best-friend: $x) isa best-friendship;}`

Here, we unify `$r` with `$_0` (anonymous variable for relationship instance in the rule `when`), and unify `$_friendship` (a labeled variable) to `$_best-friendship` (corresponding labelled variable). To make a valid answer, `$_best-friendship` must be compatible with the requirements of `$_friendship` from the query - ie. that it be of type `friendship`, or any of its subtypes (a `best-friendship` would in fact satisfy the Requirements).

Note: this also requires that a Unifier now computes a mapping between `Identifier`, rather than `Reference.Name`, and must be the output of any rule `then` materialisation.

Note: `Requirements` can later be replaced by query rewriting, if we proceed carefully and deal with all the edge cases of rewriting.

## What are the changes implemented in this PR?
* Unifiers now have a component called `Requirements` that can be used to filter out answers during reverse unification
* Materialisation and unification now operate over `Map<Identifier, Concept>`
* leads to: updating `AnswerState` to make the flow clearer: `AnswerState.UpstreamVars.Initial` -> `AnswerState.DownstreamVars.Mapped/Unified` -> `Answerstate.UpstreamVars.Derived`
* delete `//logic/transformer`: move `//logic/transformer/Unifier` -> `//logic/resolvable/Unifier` and `//logic/transformer/Mapping` -> `//reasoner/resolution/answer/Mapping`
* refactor and extend unification tests to more closely follow how unification should look in the execution flow of reasoning
* Minor bug fixes and architectural changes in `resolver`s