Suppress session timeout while transaction is open (#5986)
## What is the goal of this PR?

Grakn times out inactive sessions after a short period. Clients are supposed to send pulses at regular intervals to ensure sessions do not get timed out. But if a client is using all its available cores to do work in a transaction, it may be unable to send pulses. We now mitigate that by suppressing session timeout while any Transaction is open on that Session.

## What are the changes implemented in this PR?

- Session timeout is now suppressed while any Transaction is open on that session