Permit override of played roles in Plays constraints (#6050)
## What is the goal of this PR?

Specifying the scope of an overridden role is not allowed syntactically when specifying a Plays constraint in graql, and previously, it was not possible to infer a label appropriately for overridden roles. We added, in combination with some changes in graql, the ability to correctly infer the scope of the overridden type, allowing overridden types in definitions.

## What are the changes implemented in this PR?

Where an overridden type is supplied in a define query's plays clause, this is caught in definePlays and the overriding role's supertypes are scanned for an appropriately named roleType. If one is found, this roletype's relation is used as a scope for the overridden type, allowing it to be correctly added to the Grakn graph.