Rewritten Producers to be thread-safe with a clean algorithm (#6052)
## What is the goal of this PR?

The first draft of the Async Producer, that came out with 2.0.0-alpha release, was not yet fully thread-safe, which cause plenty of bugs that were all caused by various race conditions. It also caused "thread leaks" that overburdened the CPU to the point of unusable over time. We have now fixed this problem by rewriting the Async Producer to be thread-safe from the start, with an algorithm that is much easier to follow, which allows us to easily improve it in the future.

## What are the changes implemented in this PR?

Fixes #6042