Reasoner planning (#6051)
## What is the goal of this PR?

We add a rudimentary planner for the order of resolution by the reasoner. A query is split into resolvable parts: those that can be retrieved those that can be concluded. These resolvables can have inter-dependencies that need to be observed for correctness, which the planner accounts for. Within these bounds the planner uses basic heuristics to choose the plan, for example: resolving the concludable that can trigger the fewest rules first.

## What are the changes implemented in this PR?

- Implement the reasoner planner. The order of Resolvable prioritisation is:
  1. A Retrievable with satisfied dependencies, with the most as-yet-unplanned variables.
  2. A Concludable with satisfied dependencies, with the least applicable rules.
  3. A disconnected Retrievable, with satisfied dependencies, with the most as-yet-unplanned variables.
  4. A disconnected Concludable, without it's dependencies satisfied, which can occur with co-dependent concludables (will require reasoner reiteration).
- Rename `ResolverRegistry` to `ResolverManager`, which now handles the planning of resolvables.