Fix a concurrency bug in async producer which may result in illegal state exception (#6067)
## What is the goal of this PR?

Fix a concurrency bug in async producer which may result in illegal state exception.

## What are the changes implemented in this PR?

Currently there's check in the async job on if `runningJobs` contains the iterator. It's possible that this job is started before we put the future into the job. Therefore we always add `CompletableFuture` to the `runningJobs` before we start the job.

Fix https://github.com/graknlabs/grakn/issues/6066

Fix https://github.com/graknlabs/grakn/issues/6061