Add prefetch option for queries, which defaults to false for inserts (#6077)
## What is the goal of this PR?

On executing a Graql query, Grakn automatically streams the first batch of responses back to the client. This ensures they get their answers to a `match` query as fast as possible. But, for an `insert` query, they usually don't need those answers. And the CPU cost of compiling these unwanted response messages was inflicting a significant performance penalty. To remedy this, we made `prefetch` a configurable per-query option, which defaults to true, except for Insert queries where it defaults to false.

## What are the changes implemented in this PR?

- Don't prefetch responses to Insert queries unless specified with a configurable flag in QueryOptions (fixes https://github.com/graknlabs/grakn/issues/6072)