Optimise key ownership validation at commit time (#6099)
## What is the goal of this PR?
We make key validation much cheaper, when no keys can be present. In our previous implementation of concept validation, we checked whether required key attributes are owned - however in practice we accidentally scanned ALL attributes owned if _no_ attributes are owned as keys! This PR changes the implementation to only check for keys if there are key-attributes.

## What are the changes implemented in this PR?
* Only scan owned attributes by keyed type, if the attribute is required to own any keys.