Validate rules may only infer insertable type combinations (#6048)
## What is the goal of this PR?

We use `TypeResolver` to validate rules before insertion. First, we check that removing the `Then`'s types that cannot appear in the `when` does not make the `then` unsatisfiable. Next, we check to see if the rule's `when` has type permutations that the `then` cannot satisfy. If there is, we invalidate the rule and throw an error.

## What are the changes implemented in this PR?

* Adjusted `TypeResolver` to have an optional flag to check specify if the `conjunction` it is running on also needs to be insertable.
* Added new validation logic in `Rule` class to check if the rule's `then` and `when` resolved types are compatible.
* Added `isSystemRefernce()` helper method to `SystemReference`.