Throw exception correctly on transaction lock timeout (#6111)
## What is the goal of this PR?

When a transaction fails to acquire a write lock, presently, it throws `ERROR: Error: 13 INTERNAL: null Please check server logs for the stack trace.` when a commit is attempted, which is incorrect. We made it instead throw  `ERROR: Error: 13 INTERNAL: [TXN13] Invalid Transaction Operation: Could not acquire lock for data transaction. A schema session may have been left open. Please check server logs for the stack trace.` immediately, which is much more informative.


## What are the changes implemented in this PR?

Check the return value from trying to acquire a lock and correctly throw an exception if the result is 0, meaning the attempt timed out.