Prevent losing disjunctions' conjunctions (#6116)
## What is the goal of this PR?
We currently do an unintuitive equality check for Conjunctions, depending only on the identifier of the variables - regardless of constraints, which can lead to losing "equal" conjunctions. To prevent deduplicating conjunctions we create in a disjunction, we store them in a list rather than a set. This can be correctly implemented after is https://github.com/graknlabs/grakn/issues/6115 is complete.

## What are the changes implemented in this PR?
* Store conjunctions in a list, within a disjunction
* Bump `@graknlabs_behaviour`, which contains a test that fails for this case without this fix.