Resolution and reiteration test (#6107)
## What is the goal of this PR?

Reasoner resolution relies on a message-passing actor model internally. In order to test this we have integration tests that check that the messages sent and received are as expected. We upgrade these tests (which were mocked) to be compatible with the now functional reasoner. We also create a thorough test case for reasoner reiteration, which is required in order to run Nth passes over materialised answers in order to find the full and correct set of answers for a query.

## What are the changes implemented in this PR?

- Upgrade `ResolutionTest`.
- Split our reiteration test and upgrade.
- The ReiterationTest is ignored until we fix a bug.