Fix traversal error that lead to a role being scoped twice in the same relation (#6134)
## What is the goal of this PR?
In the traversal engine, when scoping a role instance to a relation, we previously set the scope based on both the start and end of an edge. However, this can lead to scoping the same role twice (once from the previous edge's end, and once from the next edge's start), which throws an assertion error. We now only scope based on the end of an edge.

## What are the changes implemented in this PR?
* only set scope based on the `to` of an edge
* to make up for not setting the `from` of an edge, we catch the only case where this is required ahead of time: if the start vertex is scoped