Allow creation of "schema graph initialisation" transactions (#6137)
## What is the goal of this PR?

In Grakn Cluster, all transactions (without the `readAnyReplica` flag set) are verified against the replica state, and throw an error if the replica is not primary. This was causing an issue where a Cluster node would fail to restart correctly after crashing, because it would have to perform schema-graph initialisation again (in a transaction) and that would throw. We fix it by introducing a package-private method `initialisationTransaction` into `RocksTransaction.Schema`. It does essentially the same thing as `transaction`, but uses a different method from `TransactionSchemaFactory` - that distinction allows the two to behave differently in Cluster.

## What are the changes implemented in this PR?

Add `initialisationTransaction` to `RocksTransaction.Schema`.