Remove transaction schema factory reference (#6174)
## What is the goal of this PR?

We have removed an unused transaction schema factory reference in RocksDatabase.

It was originally added in order to allow for differentiating "transaction for schema initialisation" and "transaction for regular use". We've simplified the design since and the ability to construct "transaction for schema initialisation" and "transaction for regular use" has been moved into RocksSession.Schema. Therefore, this reference should be removed.

## What are the changes implemented in this PR?

- Remove an unused transaction schema factory reference