Resolution tracing (#6186)
## What is the goal of this PR?

We introduce a visual reasoner tracing tool that displays the internal actions of the reasoning engine. It maps out the messages passed between the actors of the message passing system that implements reasoning in Grakn. This has been built for internal use, but is exposed for use by anyone using an option on transactions.

## What are the changes implemented in this PR?

- Record messages sent between actors in the reasoner.
- Save these messages to `.dot` files, one for each request that is sent into the system.
- These `.dot` files can be used to generate `.svg` visuals using `graphviz`
- Due to the async nature of the system, requests need to be serialised to give clear traces, so `parallel` needs to be set to `false` on the transaction.
- This tracer needs to be enabled by setting `traceInference` on the transaction. This is separate from the database's log level.
- Requires https://github.com/graknlabs/protocol/pull/122