Fix rule satisfiability validation (#6231)
## What is the goal of this PR?
We previously checked that a rule is coherent/satisfiability by checking each variable in the `when` and the `then`. However, we no longer set the satisfiability of a variable at all, instead setting only `isCoherent` at the conjunction. This PR updates Rule validation to reflect this change and removes `isSatisfiable` from `Variable`

Fixes #6227 

## What are the changes implemented in this PR?
* Fix rule validation
* Add a new test case to test `when` validation, extending the current `then` validation
* Remove `isSatisfiable` getters and setters from `Variable`