Prevent rare segfaults with storage read-write lock (#6233)
## What is the goal of this PR?

To prevent segfaults that can occur when a transaction has asynchronous work calling `put()` or `get()` on storage, after another thread has called `close()` on storage. To prevent this, we insert a read-write lock that needs to be obtained before each storage operation. Benchmark simulations show this isn't a major performance bottleneck, but we should still move towards more efficient implementations if possible (starting with investigating the use of Optimistic locking for low-contention locks)


## What are the changes implemented in this PR?
* Add read-write lock to every storage object
* any data `get()` and `put()` must obtain the read lock to prevent race conditions with `close()`
* any data delete must obtain write lock to prevent race conditions with `close()` and `put()` and `get()`